requirements.txt
ccxt
openai
pandas
schedule
python-dotenv
requests
urllib3
flask
flask-cors
psutil
pytz